# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger for the Horkruxes API",
        "title": "Horkruxes API",
        "contact": {},
        "license": {
            "name": "AGPLv3",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/all": {
            "get": {
                "description": "get string by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Show all messages (original messages and comments) without any more structure",
                "operationId": "all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Message"
                            }
                        }
                    }
                }
            }
        },
        "/comments/{uuid}": {
            "get": {
                "description": "get string by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Show comments to a specific message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of original message corresponding to comments",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Message"
                            }
                        }
                    }
                }
            }
        },
        "/message": {
            "get": {
                "description": "get string by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Show all original messages (for the front page)",
                "operationId": "messages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Message"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Post a new message",
                "produces": [
                    "application/json"
                ],
                "summary": "Post a new message",
                "parameters": [
                    {
                        "description": "Message",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Message"
                        }
                    }
                ]
            }
        },
        "/message/{uuid}": {
            "get": {
                "description": "get string by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Show a specific message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Message"
                        }
                    }
                }
            }
        },
        "/user/{pubkey}": {
            "get": {
                "description": "get string by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Show messages of a specific user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author ed25519 public key",
                        "name": "pubkey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Message"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Message": {
            "type": "object",
            "properties": {
                "answer-to": {
                    "description": "Used if the message is a comment to a publication",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "description": "Stored and generated",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "pod": {
                    "description": "Not saved in db but tell where it is sent from so remains in JSON",
                    "type": "string"
                },
                "public-key": {
                    "description": "Encoded in URL-safe Base64",
                    "type": "string"
                },
                "signature": {
                    "description": "Encoded in URL-safe Base64",
                    "type": "string"
                },
                "title": {
                    "description": "Stored and given by user",
                    "type": "string"
                }
            }
        }
    }
}